Partially rewrite syntaxhighlighter for more perf and speed
